requirements.txt
tensorflow
numpy
opencv-python
matplotlib
scikit-learn
requests
Pillow
pyspark
pandas
seaborn
